requirements.txt
requests-aws4auth
elasticsearch
elasticsearch-curator
python-dotenv
